requirements.txt
Flask==3.0.3
PyPDF2==3.0.1
gunicorn
Pillow==10.4.0
pdf2docx==0.5.6
PyMuPDF==1.26.6